Moved routes in routes folder, out of main server
