+++ client/sendPacket.js
@@ -1,0 +1,5 @@
+const sendPacket = (success, message, content={}) => {
+  return { success: success, message: message, content: content };
+}
+
+module.exports = sendPacket;
