+++ client/server.js
@@ -3,27 +3,19 @@ if (process.env.NODE_ENV != 'production') {
 }
 
 const express = require('express');
+const pino = require('express-pino-logger')();
+
 const bodyParser = require('body-parser');
 const expressSession = require('express-session');
-const path = require('path');
-const pino = require('express-pino-logger')();
-const fs = require('fs');
 const passport = require('passport');
 require('./passport-config');
-const bcrypt = require('bcrypt');
-
-// const { mongoose } = require('./mongoConfig');
-const { mongoose, User, InvertedIndex, Crawler, Query } = require('./mongoConfig');
-
-const log = require('./logger');
-const makeRequest = require('./makeRequest');
 
 const app = express();
 
 const port = process.env.PORT || 8000;
 const root = require('path').join(__dirname, 'frontend', 'build');
 
-const maxHistoryLength = 100;
+const log = require('./logger');
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -41,155 +33,14 @@ app.use(pino);
 app.use(express.static(root));
 app.use(passport.initialize());
 app.use(passport.session());
-require('./authRoutes')(app);
 
-//Insert server response functions here
+require('./routes/auth')(app);
+require('./routes/user')(app);
+require('./routes/search')(app);
 
-app.get('/search/:query', async (request, response) => {
-  const query = request.params['query'];
-  log('query', query);
-  const data = await makeRequest('ranker', `query/${query}`);
-
-  const queryDBObject = new Query({query: query, userid: request.user ? request.user._id : null });
-  queryDBObject.save((err) => {
-    if(err) log('error', 'Could not successfully save query');
-  })
-
-  log('ranker', data.message);
-  return response.json(
-    sendPacket(data.success, `Successfully received response from ranker: ${query}`, data.content)
-  );
-});
-
-app.get('/logout', (req, res) => {
-  req.logout();
-  res.redirect("/login");
-})
-
-app.get('/checkAuthenticated', (req, res) => {
-  if(checkAuthentication(req)) return res.json(sendPacket(1, 'User is authenticated'));
-  else return res.json(sendPacket(0, 'User not authenticated'));
-})
-
-app.post('/fetchDocuments', async (req, res) => {
-  const startTime = Date.now()
-  log('fetch', 'Fetching documents from ranker')
-
-  const docUrls = req.body['docUrls'].slice(0, 60)
-  const data = await makeRequest('ranker', 'fetchDocuments', 'POST', {docUrls: docUrls})
-  const documents = data['content']['documents']
-  log('fetch', `Received documents from ranker. Execution time ${(Date.now() - startTime) / 1000} seconds.`)
-  
-  if(req.user != undefined) {
-    User.findById(req.user._id, (err, user) => {
-      if(err) log('error', 'Unable to find user')
-      else {
-        return res.json(sendPacket(1, 'Successfully fetched documents', {documents: documents, likes: user['likes'], dislikes: user['dislikes']}))
-      }
-    })
-  }
-  else return res.json(sendPacket(1, 'Successfully fetched documents', {documents: documents, likes: {}, dislikes: {}}))
-})
-
-app.get('/updateHistory', async (req, res) => {
-  const redirectUrl = req.query.redirect
-  const user = await User.findOne({ _id: req.user._id })
-
-  if (!('history' in user)) user['history']=[];
-  if (user.history.length >= maxHistoryLength) user.history.shift();
-  
-  user.history.push(redirectUrl);
-  await user.save(err => {
-    if (err) log("error", 'Error saving user history update');
-  })
-
-  if (user!==null) log("redirect", `Sending user ${user.userid} to ${redirectUrl}`)
-  return res.redirect(redirectUrl)
-})
-
-app.post('/changeLikeStatus', (req, res) => {
-  const { likeStatus, url } = req.body;
-  if(req.user === undefined) return res.json(sendPacket(0, 'Unable to save because User not logged in.'));
-  User.findById(req.user._id, (err, user) => {
-    if(err){
-      log('error', 'Error finding user. Could not change like status');
-      return res.json(sendPacket(0, 'Unable to find user'));
-    }
-
-    if(!('likes' in user)) user['likes']= {};
-    if(!('dislikes' in user)) user['dislikes']= {};
-
-    const dotReplacedUrl = url.replace(/\./g, '%114')
-
-    if (likeStatus === 1) {
-      if (dotReplacedUrl in user['dislikes']) delete user['dislikes'][dotReplacedUrl];
-      user['likes'][dotReplacedUrl]=true;
-    }
-
-    else if (likeStatus === -1) {
-      if (dotReplacedUrl in user['likes']) delete user['likes'][dotReplacedUrl];
-      user['dislikes'][dotReplacedUrl]=true;
-    }
-
-    else {
-      if (dotReplacedUrl in user['likes']) delete user['likes'][dotReplacedUrl];
-      if (dotReplacedUrl in user['dislikes']) delete user['dislikes'][dotReplacedUrl];
-    }
-
-    user.markModified('likes');
-    user.markModified('dislikes');
-
-    user.save( (err, newUser) => {
-      if(err) {
-        return res.json(sendPacket(0, 'Unable to save like status updates'));
-      }
-      return res.json(sendPacket(1, 'Document like status successfully changed', {newLikeStatus: likeStatus, user: newUser}));
-    })
-  })
-})
-
-app.get('/user', (req, res) => {
-  if(req.user === undefined) return res.json(sendPacket(0, 'Unable to save because User not logged in.'));
-  User.findById(req.user._id, (err, user) => {
-    if(err) return res.json(sendPacket(0, 'Unable to find user'))
-    return res.json(sendPacket(1, 'Successfully found user', { history: user['history'], likes: user['likes'], dislikes: user['dislikes'] }));
-  })
-})
-
-app.get('/test', (req, res) => {
-  Query.find((err, docs) => {
-    if(err) return res.json({error: err})
-    else return res.json({query: docs})
-  })
-  // Query.findOne({query: 'chicken sauce'}, (err, query) => {
-  //   if(err) return res.json({error: err})
-  //   else return res.json({query: query})
-  // })
-  // User.find({}, (err, users) => {
-  //   if(err) {
-  //     console.log("Error:", err)
-  //     return res.json({message: 'No!'})
-  //   }
-  //   else {
-  //     console.log("Found users")
-  //     console.log(users)
-  //     return res.json({message: 'Yes!'})
-  //   }
-  // })
-})
-
-app.get('*', (request, response) => {
+app.get('*', (_, response) => {
   return response.sendFile('index.html', { root });
 });
-
-function checkAuthentication(request) {
-  return request.isAuthenticated() && request.user; // == some stored value
-}
-
-function sendPacket(success, message, content = {}) {
-  return { success: success, message: message, content: content };
-}
-//Insert helper functions here
 
 app.listen(port, () => {
   log('info', `Client server is listening on port ${port}.`);
